requirements.txt
tensorflow>=2.19.0
opencv-python>=4.12.0
numpy>=2.1.3
matplotlib>=3.8.2
scikit-learn>=1.7.1
streamlit>=1.47.0
pytesseract>=0.3.13
seaborn>=0.13.2
kagglehub>=0.3.12
Pillow>=10.0.0

